rename gather to adapter
