+++ allure-adapter-plugin/build.gradle.kts
@@ -1,0 +1,31 @@
+plugins {
+    id("allure-gradle.kotlin-dsl-published-plugin")
+}
+
+group = "io.qameta.allure.gradle.adapter"
+
+dependencies {
+    implementation(project(":allure-base-plugin"))
+    testImplementation(project(":testkit-junit4"))
+    testImplementation("org.assertj:assertj-core:_")
+}
+
+tasks.test {
+    // Treat test task out-of-date if src/it changes
+    inputs.dir(layout.projectDirectory.dir("src/it")).optional()
+}
+
+gradlePlugin {
+    plugins {
+        create("allureAdapterBasePlugin") {
+            id = "io.qameta.allure-adapter-base"
+            description = "Declares common configurations for producing and consuming Allure results and reports"
+            implementationClass = "io.qameta.allure.gradle.adapter.AllureAdapterBasePlugin"
+        }
+        create("allureAdapterPlugin") {
+            id = "io.qameta.allure-adapter"
+            description = "Implements autoconfiguration for collecting data for Allure"
+            implementationClass = "io.qameta.allure.gradle.adapter.AllureAdapterPlugin"
+        }
+    }
+}
