+++ allure-adapter-plugin/src/it/adapter-all/build.gradle
@@ -1,0 +1,17 @@
+plugins {
+    id("java")
+    id("io.qameta.allure-adapter")
+}
+
+repositories {
+    mavenCentral()
+}
+
+tasks.register("printAdapters") {
+    doLast {
+        buildDir.mkdirs()
+        file("$buildDir/printAdapters.txt").write(
+            allure.adapter.adapters.toList().collect { it.toString() }.sort().toString()
+        )
+    }
+}
