--- allure-gather-plugin/src/it/adapter-all/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id("java")
-    id("io.qameta.allure-gather")
-}
-
-repositories {
-    mavenCentral()
-}
-
-tasks.register("printAdapters") {
-    doLast {
-        buildDir.mkdirs()
-        file("$buildDir/printAdapters.txt").write(
-            allure.gather.adapters.toList().collect { it.toString() }.sort().toString()
-        )
-    }
-}
