+++ allure-plugin/src/it/report-multi/module1/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'io.qameta.allure-gather'
+    id 'io.qameta.allure-adapter'
 }
 
 dependencies {
