+++ allure-plugin/src/main/kotlin/io/qameta/allure/gradle/allure/AllurePlugin.kt
@@ -1,13 +1,13 @@
 package io.qameta.allure.gradle.allure
 
-import io.qameta.allure.gradle.gather.AllureGatherPlugin
+import io.qameta.allure.gradle.adapter.AllureAdapterPlugin
 import io.qameta.allure.gradle.report.AllureReportPlugin
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
 
 /**
- * This is a shortcut for [AllureGatherPlugin.PLUGIN_NAME] and [AllureReportPlugin.PLUGIN_NAME] plugins.
+ * This is a shortcut for [AllureAdapterPlugin.PLUGIN_NAME] and [AllureReportPlugin.PLUGIN_NAME] plugins.
  * If you need an aggregate report, then use [AllureAggregateReportPlugin.PLUGIN_NAME] plugin.
  */
 open class AllurePlugin : Plugin<Project> {
@@ -16,7 +16,7 @@ open class AllurePlugin : Plugin<Project> {
     }
 
     override fun apply(target: Project): Unit = target.run {
-        apply<AllureGatherPlugin>()
+        apply<AllureAdapterPlugin>()
         apply<AllureReportPlugin>()
     }
 }
