+++ sandbox/dsl-groovy/build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 allure {
-    gather {
+    adapter {
 
     }
     report {
