+++ settings.gradle.kts
@@ -38,7 +38,7 @@ includeBuild("build-logic-commons")
 includeBuild("build-logic")
 
 include("allure-base-plugin")
-include("allure-gather-plugin")
+include("allure-adapter-plugin")
 include("allure-report-plugin")
 include("allure-plugin")
 include("testkit-junit4")
